Move job creation logic to separate file
This re-creates `gitjob.go`, this time containing logic related to
creating the git job, leaving `gitjob_controller.go` with mostly
reconciler-related logic.
